ATT: move AttUuid into its own module, impl Debug
